refactor(input): move the binding tables and the fold into their own module
A pure move, with the tests from the previous commit travelling with it. The
tables, InitInput and RestoreInput go to SOURCES/INPUT_BINDINGS.{H,CPP}; the
config IO and the per-frame polling stay in INPUT.CPP. The moved text is
byte-identical to its old home, which is checkable:

  git show HEAD~1:SOURCES/INPUT.CPP > /tmp/old.CPP
  diff <(grep -v '^$' <(sed -n '29,125p;27p;136,167p' /tmp/old.CPP)) \
       <(grep -v '^$' <(sed -n '12,$p' SOURCES/INPUT_BINDINGS.CPP))

132 lines, against the 150 the plan estimated: the difference is blank lines
and a banner comment. INPUT.CPP goes from 446 lines to 313, and loses its
include of KEYBOARD_KEYS.H with them, since the only K_ name left in the file
after the tables went is one inside a comment.

The cut is along what the tests could reach without the engine. Everything that
moved is a lookup over data, and what is left needs DefFileBuffer or a keyboard
to sample. tests/input_bindings shows the difference directly: it now links
INPUT_BINDINGS.CPP, DefineInputKeys and two keyboard hooks, where before it
needed the console, the pad, the screen and the whole DEFFILE chain stubbed or
linked for a test that touches none of them.

INPUT.H includes the new header, so the eighteen files that include INPUT.H see
the tables where they did and no caller changes. Narrowing that to the handful
that actually rebind keys is step 7 of the recipe and belongs to increments 3
and 6, not to a move. DefKeysDefault95 is declared in the new header because
ReadInputConfig names it directly; MAX_INPUT and T_DEF_KEY go with the table
they size.

Two things recorded while moving them, both already true: the 1997 file carried
a DOS layout beside this one and the port kept only the Windows 95 set, and
TabInputBit is read by nothing in this build since the combined table replaced
it. Neither is changed here.

docs/CONFIG.md now points at the header MAX_INPUT lives in, and the plan records
what increment 0 produced, including the one cell where the shipped layout
departs from the 1997 table.

Verified: game builds and 63 host tests pass under gcc, clang, the demo
configuration and ASan/UBSan; check-arch clean at 561 files.
