+++ tests/input_bindings/held_keys.h
@@ -1,5 +1,5 @@
-#ifndef TESTS_INPUT_BINDINGS_ENGINE_STUBS_H
-#define TESTS_INPUT_BINDINGS_ENGINE_STUBS_H
+#ifndef TESTS_INPUT_BINDINGS_HELD_KEYS_H
+#define TESTS_INPUT_BINDINGS_HELD_KEYS_H
 
 #include <SYSTEM/ADELINE_TYPES.H>
 
@@ -8,4 +8,4 @@
 void HoldNone(void);
 void HoldOnly(U32 key);
 
-#endif // TESTS_INPUT_BINDINGS_ENGINE_STUBS_H
+#endif // TESTS_INPUT_BINDINGS_HELD_KEYS_H
